refactor: 헤더 반응형 수정
