+++ src/components/header/MainHeader.tsx
@@ -2,53 +2,14 @@
 
 import { useEffect } from "react";
 import { RootState } from "@/lib/store";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import Link from "next/link";
-import useMediaQuery from "@/hooks/useMediaQuery";
-import Image from "next/image";
-import { openModal } from "@/lib/store/modal";
-import UserMenu from "./UserMenu";
-import SearchReviews from "./SearchReviews";
+import ResponsiveSearch from "./ResponsiveSearch";
+import ResponsiveUser from "./ResponsiveUser";
 
 export default function MainHeader() {
-  const { user, isLoading } = useSelector((state: RootState) => state.user);
-  const isMobile = useMediaQuery("(max-Width:768px)");
+  const { isLoading } = useSelector((state: RootState) => state.user);
   const isOpen = useSelector((state: RootState) => state.modal.isOpen);
-  const dispatch = useDispatch();
-
-  // 리뷰 검색창
-  let searchReviews;
-
-  if (isMobile) {
-    searchReviews = (
-      <button
-        onClick={() => dispatch(openModal({ modalType: "searchReviews" }))}
-      >
-        <Image
-          src="/images/search-icon.svg"
-          alt="검색 아이콘"
-          width={40}
-          height={40}
-        />
-      </button>
-    );
-  } else {
-    searchReviews = <SearchReviews width="w-72" />;
-  }
-
-  // user 표시
-  let content;
-
-  if (user) {
-    content = <UserMenu user={user} />;
-  } else {
-    content = (
-      <>
-        <Link href="/login">로그인</Link>
-        <Link href="/signup">회원가입</Link>
-      </>
-    );
-  }
 
   useEffect(() => {
     if (isOpen) {
@@ -66,8 +27,10 @@ export default function MainHeader() {
         </Link>
         {isLoading ? null : (
           <nav className="flex items-center gap-2 text-base font-semibold sm:gap-7">
-            {searchReviews}
-            <div className="relative flex gap-7 text-blue">{content}</div>
+            <ResponsiveSearch />
+            <div className="relative flex gap-7 text-blue">
+              <ResponsiveUser />
+            </div>
           </nav>
         )}
       </div>
