+++ src/components/header/ResponsiveUser.tsx
@@ -1,0 +1,21 @@
+import Link from "next/link";
+import useMediaQuery from "@/hooks/useMediaQuery";
+import { useSelector } from "react-redux";
+import { RootState } from "@/lib/store";
+import UserMenu from "./UserMenu";
+
+export default function ResponsiveUser() {
+  const isMobile = useMediaQuery("(max-Width:768px)");
+  const { user } = useSelector((state: RootState) => state.user);
+
+  if (user) return <UserMenu user={user} />;
+
+  if (isMobile === null) return null;
+  if (isMobile) return <Link href="/login">로그인</Link>;
+  return (
+    <>
+      <Link href="/login">로그인</Link>
+      <Link href="/signup">회원가입</Link>
+    </>
+  );
+}
