+++ src/hooks/useMediaQuery.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 
 export default function useMediaQuery(query: string) {
-  const [matches, setMatches] = useState(false);
+  const [matches, setMatches] = useState<boolean | null>(null);
 
   useEffect(() => {
     const mediaQuery = window.matchMedia(query);
